+++ src/core/test-helpers/assert-either.ts
@@ -1,0 +1,8 @@
+import * as E from 'fp-ts/Either';
+
+export function assertEither<T, U>(result: E.Either<T, U>, assertion: (data: U) => void): void {
+  E.match<T, U, void>(
+    (error) => fail(error),
+    (data) => assertion(data),
+  )(result);
+}
